Fix dl-apply.sh script and regenerate faq.md page.

+++ scripts/dl-apply.sh
@@ -7,8 +7,9 @@
 URL_Apply="https://raw.githubusercontent.com/wiki/hakierspejs/wiki/Jak-si%C4%99-zapisa%C4%87.md"
 URL_FAQ="https://raw.githubusercontent.com/wiki/hakierspejs/wiki/FAQ.md"
 
-if [[ -f "./apply.md" || -f "./faq.md" ]]; then
-	rm -r "./apply.md"
+if [[ -f "./apply.md" ]] && [[ -f "./faq.md" ]]; then
+	rm -r "./apply.md" 
+	rm -r "./faq.md"
 fi
 
 # apply page
